refactor(index): decompose neo4j_store.py god-module into neo4j/ subpackage
Split the 2877-line omniscience_index.stores.neo4j_store module into a
cohesive neo4j/ subpackage without changing behaviour or public API:

- neo4j/_cypher.py  (868 lines): all Cypher/DDL string constants,
  label/key constants, and the import-time _ensure_workspace_predicate
  regression guard sweep — ACL invariant enforcement is preserved.

- neo4j/_tx.py (44 lines): low-level async transaction runner helpers
  (_run_write_stmt, _run_write_returning, _run_read_stmt).

- neo4j/mappers.py (369 lines): pure-Python row converters, metadata
  JSON coercion, datetime helpers, state-change fingerprinting,
  traversal Cypher builder, and parameter-map builders.

- neo4j/store.py (952 lines): Neo4jStoreConfig dataclass and the full
  Neo4jGraphStore class, delegating to the helper modules above.

neo4j_store.py becomes a thin re-export sheet that re-publishes every
symbol (public and private) that existing callers import, preserving
backward compatibility with no import-site changes required.

mock.patch compatibility: AsyncGraphDatabase is explicitly imported into
neo4j_store.py's namespace so patch("...neo4j_store.AsyncGraphDatabase.driver")
continues to intercept Neo4jGraphStore.__init__ calls as before.

Validation:
- uv run ruff check packages/index: clean (only pre-existing writer.py whitespace)
- python -c "from omniscience_index.stores.neo4j_store import Neo4jGraphStore": OK
- All private symbols still importable from neo4j_store path: OK
- Non-container tests: 150 passed, 8 pre-existing failures (identical to original)
- Contract tests (Docker): 4 passed, 3 pre-existing failures (identical to original)
- mypy: 1 pre-existing error (improved from 2 on original)
